[MCUX-81864][DCP] Implement AES cache handling
The DCP module uses its own DMA, and the cached data is not visible to the CPU.
Therefore, explicit cache management is required.

On other platforms, we typically configure the cache in write-through mode and
invalidate output buffers coming from hardware IPs. However, the RT10xx device
with DCP is affected by an ARM erratum related to this cache mode.

To work around this issue, we implemented a handling mechanism where all AES
data buffers are copied to aligned memory. These buffers are then cleared and
invalidated as needed to ensure data consistency.

Support for multipart HASH operations is not included in this commit,
as it would require significantly more effort and hardware resources.

Signed-off-by: Vit Solcik (nxf07484) <vit.solcik@nxp.com>