[MCUX-83007][S200] Add accelerated wrapper for psa_export_public_key()
This wrapper is only available if the appropriate firmware is loaded,
as the functionality is not available from ROM. The wrapper checks for
any loaded firmware before utilizing the API. If no FW is found, we fall
back to a PSA built-in implementation, if it is compiled in.

Functionality is also limited to SECP-R1 and Brainpool-R1 curves.

Signed-off-by: Patrik Nemeth